created findall query for clients

+++ group assign co/roadco/src/main/java/groupassign/roadco/model/repositories/ClientRepo.java
@@ -1,10 +1,13 @@
 package groupassign.roadco.model.repositories;
 
+
+
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 import groupassign.roadco.model.entities.Client;
 
 @Repository
-public interface ClientRepo extends CrudRepository<Client,Integer>{
-    
+public interface ClientRepo extends CrudRepository<Client,Long>{
+
+ 
 }
